% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{adviceNgrams}
\alias{adviceNgrams}
\title{Pre-trained advice concreteness features}
\format{
A (truncated) matrix of ngram feature counts for alignment to the pre-trained advice glmnet model.
}
\source{
Yeomans (2020). A Concrete Application of Open Science for Natural Language Processing.
}
\usage{
adviceNgrams
}
\description{
For internal use only. This dataset demonstrates the ngram features that are used for the pre-trained adviceModel.
}
\keyword{datasets}
