% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bootstrap_list}
\alias{bootstrap_list}
\title{Concreteness mTurk Word List}
\format{A list of 85,942 words where concreteness was imputed using word embeddings}
\usage{
bootstrap_list
}
\description{
Word list from Paetzlod & Specia (2017)
}
\keyword{datasets}
