% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textFuncs.R
\name{cleantext}
\alias{cleantext}
\title{Text Cleaner}
\usage{
cleantext(text, language = "english", stop.words = TRUE,
  number.words = TRUE)
}
\arguments{
\item{text}{character vector of strings to clean}

\item{language}{character language to use for cleaning (default is english)}

\item{stop.words}{logical should stop words be kept? default is TRUE}

\item{number.words}{logical should numbers be converted to words? default is TRUE}
}
\description{
background function to load
}
\keyword{internal}
