% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concDict.R
\name{concDict}
\alias{concDict}
\title{Open-Domain Concreteness Dictionaries}
\usage{
concDict(
  texts,
  wordlist = NULL,
  stop.words = TRUE,
  number.words = TRUE,
  shrink = FALSE,
  fill = FALSE,
  minwords = 0,
  num.mc.cores = 1
)
}
\arguments{
\item{texts}{character Vector of documents to classify}

\item{wordlist}{Dictionary to be used.}

\item{stop.words}{logical should stop words be kept? default is TRUE}

\item{number.words}{logical should numbers be converted to words? default is TRUE}

\item{shrink}{logical should scores on shorter documents be regularized? default is FALSE}

\item{fill}{logical Should empty cells be assigned the mean rating? Default is FALSE.}

\item{minwords}{numeric all documents with less words than this return NA. default is 0 (i.e. keep all documents)}

\item{num.mc.cores}{numeric number of cores for parallel processing - see parallel::detectCores()}
}
\value{
concreteness score for each document
}
\description{
background function to load
}
\keyword{internal}
