% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngramComponents.R
\name{ctxpand}
\alias{ctxpand}
\title{Contraction Expander}
\usage{
ctxpand(text)
}
\arguments{
\item{text}{character vector of sentences to un-contract.}
}
\value{
character Vector of sentences without contractions.
}
\description{
background function to load.
}
\keyword{internal}
