% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc2concrete.R
\name{doc2concrete}
\alias{doc2concrete}
\title{Concreteness Scores}
\usage{
doc2concrete(texts, domain = c("advice", "open"), stop.words = TRUE,
  number.words = TRUE, shrink = TRUE, length = TRUE)
}
\arguments{
\item{texts}{character A vector of texts, each of which will be tallied for concreteness.}

\item{domain}{character Indicates the domain from wihch the text data was collected (see details).}

\item{stop.words}{logical should stop words be kept? default is TRUE}

\item{number.words}{logical should numbers be converted to words? default is TRUE}

\item{shrink}{logical Should open-domain concreteness models regularize low-count words? Default is TRUE.}

\item{length}{logical Should open-domain concreteness models treat document length as a feature of concreteness? Default is TRUE}
}
\value{
A vector of concreteness scores, with one value for every item in `text`.
}
\description{
Detects linguistic markers of politeness in natural language.
    This function is the workhorse of the \code{doc2concrete} package, taking an N-length vector of text documents and returning an N-length vector of concreteness scores.
}
\examples{

\dontrun{

data("feedback_dat")

doc2concrete(feedback_dat$feedback, domain="open")

hist(doc2concrete(feedback_dat$feedback, domain="open"))

cor(doc2concrete(feedback_dat$feedback, domain="open"),feedback_dat$concrete)



}

}
