% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngramComponents.R
\name{doublestacker}
\alias{doublestacker}
\title{Doublestacker}
\usage{
doublestacker(wdcts)
}
\arguments{
\item{wdcts}{matrix Token counts that will have doubled column names condensed.}
}
\value{
Token count matrix with no doubled column names.
}
\description{
background function to load
}
\keyword{internal}
