% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dtmSet}
\alias{dtmSet}
\title{Pre-trained concreteness features}
\format{A matrix of ngram feature counts for the pre-trained glmnet model}
\source{
Yeomans (2019). Concreteness, Concretely.
}
\usage{
dtmSet
}
\description{
Pre-trained concreteness features
}
\keyword{datasets}
