% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{feedback_dat}
\alias{feedback_dat}
\title{Personal Feedback Dataset}
\format{A data frame with 171 rows and 2 variables:
\describe{
  \item{feedback}{character text of feedback from writers}
  \item{concrete}{numeric average specificity score from readers}
}}
\source{
Blunden, H., Green, P., & Gino, F. (2018).

"The Impersonal Touch: Improving Feedback-Giving with Interpersonal Distance."

Academy of Management Proceedings, 2018.
}
\usage{
feedback_dat
}
\description{
A dataset containing responses from people on Mechanical Turk, writing
feedback to a recent collaborator, that were then scored by other Turkers
for feedback specificity.
}
\keyword{datasets}
