% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{finance_list}
\alias{finance_list}
\title{Concreteness Finance Word List}
\format{
A data frame with 40,004 rows and 2 variables.
\describe{
  \item{Word}{character text of a word with an entry in this dictionary}
  \item{Conc.M}{average concreteness score for that word (from 1-5)}
}
}
\source{
Reyt & Yeomans (working paper)
}
\usage{
finance_list
}
\description{
The mTUrk-annotated list, with the 1,000 most common words re-annotated by finance professionals as an in-domain dictionary.
}
\keyword{datasets}
