% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngramComponents.R
\name{gramstem}
\alias{gramstem}
\title{Stemmer}
\usage{
gramstem(text, wstem = "all", ngrams = 1, language = "english")
}
\arguments{
\item{text}{character vector of strings to clean.}

\item{wstem}{character Which words should be stemmed? Defaults to "all".}

\item{ngrams}{numeric Vector of ngram lengths to be included. Default is 1 (i.e. unigrams only).}

\item{language}{Language for stemming. Default is "english".}
}
\value{
Sentence of stemmed words.
}
\description{
background function to load.
}
\keyword{internal}
