% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mturk_list}
\alias{mturk_list}
\title{Concreteness mTurk Word List}
\format{
A data frame with 39,954 rows and 2 variables.
\describe{
  \item{Word}{character text of a word with an entry in this dictionary}
  \item{Conc.M}{average concreteness score for that word (from 1-5)}
}
}
\source{
Brysbaert, M., Warriner, A. B., & Kuperman, V. (2014). Concreteness ratings for 40 thousand generally known English word lemmas. Behavior Research Methods, 46(3), 904-911.
}
\usage{
mturk_list
}
\description{
Word list from Brysbaert, Warriner & Kuperman (2014). A list of 39,954 words that have been hand-annotated by crowdsourced workers for concreteness.
}
\keyword{datasets}
