% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ngramList}
\alias{ngramList}
\title{Pre-trained concreteness features}
\format{A (truncated) matrix of ngram feature counts for alignment to the pre-trained glmnet model.}
\source{
Yeomans (2019). Concreteness, Concretely.
}
\usage{
ngramList
}
\description{
Pre-trained concreteness features
}
\keyword{datasets}
