% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngramComponents.R
\name{ngrammer}
\alias{ngrammer}
\title{Phraser}
\usage{
ngrammer(onewords, ngram)
}
\arguments{
\item{onewords}{character Vector of words to combine into phrases.}

\item{ngram}{numeric Length of phrases to create.}
}
\value{
Sentence of stemmed ngrams.
}
\description{
background function to load
}
\keyword{internal}
