% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngramComponents.R
\name{overlaps}
\alias{overlaps}
\title{Overlap cleaner}
\usage{
overlaps(high, low, cutoff = 1)
}
\arguments{
\item{high}{matrix Token counts that will all be kept.}

\item{low}{matrix Token counts that will evaluated (and pruned) for overlapping.}

\item{cutoff}{numeric Threshold (as cosine distance) for including overlapping tokens. Default is 1 (i.e. all tokens included).}
}
\value{
Combined token count matrix.
}
\description{
background function to load
}
\keyword{internal}
