% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{planNgrams}
\alias{planNgrams}
\title{Pre-trained plan concreteness features}
\format{
A (truncated) matrix of ngram feature counts for alignment to the pre-trained planning glmnet model.
}
\source{
Yeomans (2020). A Concrete Application of Open Science for Natural Language Processing.
}
\usage{
planNgrams
}
\description{
For internal use only. This dataset demonstrates the ngram features that are used for the pre-trained planModel.
}
\keyword{datasets}
