% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngramComponents.R
\name{stemexcept}
\alias{stemexcept}
\title{Conditional Stemmer}
\usage{
stemexcept(sentence, excepts, language = "english")
}
\arguments{
\item{sentence}{character Vector of sentences to stem.}

\item{excepts}{character Vector of words that should not be stemmed.}

\item{language}{Language for stemming. Default is "english".}
}
\value{
Sentence of stemmed words.
}
\description{
background function to load
}
\keyword{internal}
