% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngramComponents.R
\name{stemmer}
\alias{stemmer}
\title{Stemmer}
\usage{
stemmer(text, wstem = "all", language = "english")
}
\arguments{
\item{text}{character vector of strings to clean.}

\item{wstem}{character Which words should be stemmed? Defaults to "all".}

\item{language}{Language for stemming. Default is "english".}
}
\value{
Sentence of stemmed words.
}
\description{
background function to load.
}
\keyword{internal}
