% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textFuncs.R
\name{textcounter}
\alias{textcounter}
\title{Text Counter}
\usage{
textcounter(counted, texts, fixed = TRUE)
}
\arguments{
\item{counted}{vector of character strings to count}

\item{texts}{vector of character strings (i.e. documents to classify)}

\item{fixed}{should regular expressions be ignored? default is TRUE}
}
\value{
vector with same length as texts, counting all instances of counted
}
\description{
background function to load
}
\keyword{internal}
