% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngramComponents.R
\name{textformat}
\alias{textformat}
\title{Text Formatter}
\usage{
textformat(text, punct = FALSE)
}
\arguments{
\item{text}{character Vector of strings to clean.}

\item{punct}{logical Should punctuation be kept as tokens? Default is TRUE.}
}
\value{
character Vector of cleaned strings.
}
\description{
background function to load.
}
\keyword{internal}
