% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngramComponents.R
\name{vocabmatch}
\alias{vocabmatch}
\title{Feature Count Matcher}
\usage{
vocabmatch(hole, peg)
}
\arguments{
\item{hole}{matrix Token counts in model data.}

\item{peg}{matrix Token counts in new data.}
}
\value{
Token counts matrix from new data, with column names that match the model data.
}
\description{
background function to load
}
\keyword{internal}
