% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textFuncs.R
\name{word_list}
\alias{word_list}
\title{Word List Counter}
\usage{
word_list(text, wordlist = NULL, stop.words = TRUE,
  number.words = TRUE, minwords = 0)
}
\arguments{
\item{text}{character string (i.e. document to classify)}

\item{wordlist}{dictionary to be used}

\item{stop.words}{logical should stop words be kept? default is TRUE}

\item{number.words}{logical should numbers be converted to words? default is TRUE}

\item{min.words}{numeric all documents with less words than this return NA. default is 0 (i.e. keep all documents)}
}
\value{
variance-weighted log odds ratio of prevalence across samples
}
\description{
background function to load
}
\keyword{internal}
