% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R, R/domainModels.R
\docType{data}
\name{adviceModel}
\alias{adviceModel}
\title{Pre-trained Concreteness Detection Model for Advice}
\format{
A pre-trained glmnet model
}
\source{
Yeomans (2020). A Concrete Application of Open Science for Natural Language Processing.
}
\usage{
adviceModel

adviceModel(texts, num.mc.cores = 1)
}
\arguments{
\item{texts}{character A vector of texts, each of which will be tallied for concreteness.}

\item{num.mc.cores}{numeric number of cores for parallel processing - see parallel::detectCores(). Default is 1.}
}
\value{
numeric Vector of concreteness ratings.
}
\description{
This model was pre-trained on 3289 examples of feedback on different tasks (e.g. writing a cover letter, boggle, workplace annual reviews). All of those documents were annotated by research assistants for concreteness, and this model simulates those annotations on new documents.

Model pre-trained on advice data.
}
\keyword{datasets}
\keyword{internal}
