% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{adviceModel}
\alias{adviceModel}
\title{Pre=trained Concreteness Detection Model for Advice}
\format{A pre-trained glmnet model}
\source{
Yeomans (2019). Concreteness, Concretely.
}
\usage{
adviceModel
}
\description{
Trained on advice datasets
}
\keyword{datasets}
