% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{adviceNgrams}
\alias{adviceNgrams}
\title{Pre-trained advice concreteness features}
\format{A (truncated) matrix of ngram feature counts for alignment to the pre-trained advice glmnet model.}
\source{
Yeomans (2019). Concreteness, Concretely.
}
\usage{
adviceNgrams
}
\description{
Pre-trained advice concreteness features
}
\keyword{datasets}
