% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bootstrap_list}
\alias{bootstrap_list}
\title{Concreteness mTurk Word List}
\format{
A data frame with 85,942 rows and 2 variables.
\describe{
  \item{Word}{character text of a word with an entry in this dictionary}
  \item{Conc.M}{predicted concreteness score for that word (from 100-700)}
}
}
\source{
#' Paetzold, G., & Specia, L. (2016, June). Inferring psycholinguistic properties of words. In Proceedings of the 2016 Conference of the North American Chapter of the Association for Computational Linguistics: Human Language Technologies (pp. 435-440).
}
\usage{
bootstrap_list
}
\description{
Word list from Paetzold & Specia (2016). A list of 85,942 words where concreteness was imputed using word embeddings.
}
\keyword{datasets}
