% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bootstrap_list}
\alias{bootstrap_list}
\title{Concreteness mTurk Word List}
\format{
A list of 85,942 words where concreteness was imputed using word embeddings
}
\source{
#' Paetzold, G., & Specia, L. (2016, June). Inferring psycholinguistic properties of words. In Proceedings of the 2016 Conference of the North American Chapter of the Association for Computational Linguistics: Human Language Technologies (pp. 435-440).
}
\usage{
bootstrap_list
}
\description{
Word list from Paetzlod & Specia (2017)
}
\keyword{datasets}
