% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngramComponents.R
\name{cleantext}
\alias{cleantext}
\title{Text Cleaner}
\usage{
cleantext(text, language = "english", punct = FALSE,
  stop.words = TRUE, number.words = TRUE)
}
\arguments{
\item{text}{character Vector of strings to clean.}

\item{language}{character Language to use for cleaning. Default is "english".}

\item{punct}{logical Should punctuation be kept as tokens? Default is TRUE.}

\item{stop.words}{logical Should stop words be kept? default is TRUE.}

\item{number.words}{logical Should numbers be converted to words? default is TRUE.}
}
\value{
character Vector of cleaned strings.
}
\description{
background function to load.
}
\keyword{internal}
