% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concDict.R
\name{concDict}
\alias{concDict}
\title{Open-Domain Concreteness Dictionaries}
\usage{
concDict(
  texts,
  wordlist = NULL,
  shrink = FALSE,
  stop.words = TRUE,
  number.words = TRUE,
  minwords = 0,
  num.mc.cores = 1
)
}
\arguments{
\item{wordlist}{Dictionary to be used.}

\item{shrink}{logical should scores on shorter documents be regularized? default is FALSE}

\item{stop.words}{logical should stop words be kept? default is TRUE}

\item{number.words}{logical should numbers be converted to words? default is TRUE}

\item{minwords}{numeric all documents with less words than this return NA. default is 0 (i.e. keep all documents)}

\item{num.mc.cores}{numeric number of cores for parallel processing - see parallel::detectCores()}

\item{text}{character Vector of documents to classify}
}
\value{
variance-weighted log odds ratio of prevalence across samples
}
\description{
background function to load
}
\keyword{internal}
