% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textFuncs.R
\name{ctxpand}
\alias{ctxpand}
\title{Contraction Expander}
\usage{
ctxpand(text)
}
\arguments{
\item{text}{character vector of strings to un-contract}
}
\description{
background function to load
}
\keyword{internal}
