% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mturk_list}
\alias{mturk_list}
\title{Concreteness mTurk Word List}
\format{A list of 39,954 words that have been hand-annotated for concreteness}
\source{
Brysbaert, M., Warriner, A. B., & Kuperman, V. (2014). Concreteness ratings for 40 thousand generally known English word lemmas. Behavior Research Methods, 46(3), 904-911.
}
\usage{
mturk_list
}
\description{
Word list from Brysbaert, Warriner & Kuperman (2014)
}
\keyword{datasets}
