% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R, R/domainModels.R
\docType{data}
\name{planModel}
\alias{planModel}
\title{Pre-trained Concreteness Detection Model for Plan-Making}
\format{
A pre-trained glmnet model
}
\source{
Yeomans (2020). A Concrete Application of Open Science for Natural Language Processing.
}
\usage{
planModel

planModel(texts, num.mc.cores = 1)
}
\arguments{
\item{texts}{character A vector of texts, each of which will be tallied for concreteness.}

\item{num.mc.cores}{numeric number of cores for parallel processing - see parallel::detectCores(). Default is 1.}
}
\value{
numeric Vector of concreteness ratings.
}
\description{
This model was pre-trained on 5,172 examples of pre-course plans from online courses at HarvardX. Each plan was annotated by research assistants for concreteness, and this model simulates those annotations on new plans.

Model pre-trained on planning data.
}
\keyword{datasets}
\keyword{internal}
