% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{planModel}
\alias{planModel}
\title{Pre=trained Concreteness Detection Model for Plan-Making}
\format{A pre-trained glmnet model}
\source{
Yeomans (2019). Concreteness, Concretely.
}
\usage{
planModel
}
\description{
Trained on planning datasets
}
\keyword{datasets}
