% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{planNgrams}
\alias{planNgrams}
\title{Pre-trained plan concreteness features}
\format{A (truncated) matrix of ngram feature counts for alignment to the pre-trained planning glmnet model.}
\source{
Yeomans (2019). Concreteness, Concretely.
}
\usage{
planNgrams
}
\description{
Pre-trained plan concreteness features
}
\keyword{datasets}
